+++ src-tauri/src/node_adapter.rs
@@ -48,7 +48,7 @@ use tokio::sync::watch;
 use tokio::time::timeout;
 
 #[cfg(target_os = "windows")]
-use crate::utils::setup_utils::setup_utils::add_firewall_rule;
+use crate::utils::windows_setup_utils::add_firewall_rule;
 
 const LOG_TARGET: &str = "tari::universe::minotari_node_adapter";
