+++ src-tauri/src/utils/mod.rs
@@ -25,5 +25,7 @@ pub mod file_utils;
 pub mod logging_utils;
 pub mod macos_utils;
 pub mod platform_utils;
-pub mod setup_utils;
+
 pub mod shutdown_utils;
+#[cfg(windows)]
+pub mod windows_setup_utils;
